Add Functionality to Create Doc Pool
Certain operations like REMOVE need documents to exist in the database
before they can be tested. This commit adds functionality for that to
happen.

It estimates how many documents the Remove operations will need (750 *
number of REMOVE operations in all test runs * horizontal scaling *
number of seconds in runtime) and creates them in a seperate collection.

This is so far untested and might take an extremely long time. It will
definatly take longer than the runtime you have set so be wary.

Change-Id: Ifd0a922e56c5a8a28a3dc36a540dbc816ca4ca1f