Refactor Test Bounding
Tests used to be bounded by runtime, i.e. run for x seconds. Now they
are bounded by number of doc operations in each run. This change was
made because to allow remove tests to be bounded by runtime we had to
predict how many douments the test would need whic was not ideal.

As a result of this it was found that the Java performer was incorrectly
multithreaded and needed some extra synchronisation.

Also it used to be that every time the performer finished a number of
operations it would send over the time taken to do that number. Now the
performer sends over how long it took to do one operation. This comes
with a lot more over head on the performer side which may affect
throughput numbers. This does not matter too much though as all tests
will suffer the same hit to throughput.

Change-Id: I07493c2aa7d679cbf41faa8ef8f975b085a21fd4