Driver Message Fix
The driver was not able to actually take two different commands in the
same .yaml file due to the way it was implemented. This has now been
fixed.

Also the Driver would make a connection to the cluster for the docPool
even if it wasn't planning to add any documents. this has been fixed.

Change-Id: I566e8d71eb4505b4f16865fe69c408a8b22470fd