Add Replace to Driver and Java Performer
Also updated GET's to use the doc pool.

Currently it is possible to have a multi operation workload but only if
there is one of each operation and remove is last in the list of
operations to do. This is because for now most of the tests should just
be single operations.

Change-Id: I6891a3512b1e75b7a7ffeb3a9c8d6158ef2858e1