Beginning of Dockerfile
Started to add a dockerfile along with some other minute changes. The
dockerfile does not currently work as Python sdk3.x and 4.x get imports
from different libraries.

Putting the docker stuff on hold for now and starting to look at getting
resource metrics such as memory (hence the reason for this commit)

Change-Id: I2aca79d4d3a2e616577e6bd4f1aa66215858020d